fixed lightref

+++ Music/LightHeightRef.py
@@ -96,9 +96,9 @@ while stream.is_active():
         #color = wheel(int (levelcolor / 240.0 * 255.0))
         for pixel in  range(levelcolor + 1, LightInfo.LED_COUNT):
             strip.setPixelColor(pixel,Color(0,0,0))
-        for pixel in  range(levelcolor / 2.0):
+        for pixel in  range(int(levelcolor / 2.0)):
             strip.setPixelColor(pixel,color)
-        for pixel in  range(LightInfo.LED_COUNT, LightInfo.LED_COUNT - levelcolor / 2.0, -1):
+        for pixel in  range(LightInfo.LED_COUNT,int(LightInfo.LED_COUNT - levelcolor / 2.0), -1):
             strip.setPixelColor(pixel,color)
         strip.show()
     time.sleep(30.0/1000)
